Moving the valkey_strtod.* files back to deps folder.
