+++ src/resp_parser.c
@@ -58,7 +58,7 @@
 #include "resp_parser.h"
 #include "server.h"
 
-#include "fast_float/valkey_strtod.h"
+#include "../deps/fast_float_c_interface/valkey_strtod.h"
 
 static int parseBulk(ReplyParser *parser, void *p_ctx) {
     const char *proto = parser->curr_location;
